Move api validation errors code to lib
